refactor: deposit files
