+++ apps/bridge-dapp/.eslintrc.json
@@ -17,6 +17,7 @@
     }
   ],
   "rules": {
-    "no-unused-vars": "warn"
+    "no-unused-vars": "off",
+    "@typescript-eslint/no-unused-vars": "error"
   }
 }
